medina-back/requirements.txt
azure-functions
openai
psycopg2-binary
requests
python-multipart